+++ CHANGELOG.md
@@ -16,12 +16,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added growth/shrink to Power Effect Builder.  
 - Added sloth/speed to Power Effect Builder. 
 - Added Beast Friend to Power Effect Builder.   
+- Added Invisibility to Power Effect Builder.  
 - Localisation for:
 - - Fear Table function.
 - - Mark Dead function.
 - Various new localisation strings.
 - New setting to check single reload by default.
-- Make tokens invisible (no player owner) or half transparent (player owner) whne the invisibility condition is applied.
+- Make tokens invisible (no player owner) or half transparent (player owner) when the invisibility condition is applied.
 ### Changed
 - Incapacitation and Bleeding Out are added as overlays instead of small icons now. Inc. will be set as small icon if Bleeding Out is applied.
 ### Fixed